Add self reflection storage feature
