+++ src/main/java/wellnus/command/HelpCommand.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import wellnus.atomichabit.feature.AtomicHabitManager;
 import wellnus.common.MainManager;
 import wellnus.exception.BadCommandException;
-import wellnus.reflection.ReflectionManager;
+import wellnus.reflection.feature.ReflectionManager;
 import wellnus.ui.TextUi;
 
 /**
