+++ src/main/java/wellnus/reflection/feature/ReflectionManager.java
@@ -1,4 +1,4 @@
-package wellnus.reflection;
+package wellnus.reflection.feature;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -10,6 +10,10 @@ import java.util.logging.Logger;
 
 import wellnus.exception.BadCommandException;
 import wellnus.manager.Manager;
+import wellnus.reflection.command.FavoriteCommand;
+import wellnus.reflection.command.GetCommand;
+import wellnus.reflection.command.HomeCommand;
+import wellnus.reflection.command.LikeCommand;
 
 /**
  * The manager for self reflection section.<br/>
@@ -48,8 +52,6 @@ public class ReflectionManager extends Manager {
             + "and improve on yourself!!";
     private static final int EMPTY_COMMAND = 0;
     private static final boolean IS_EXIT_INITIAL = false;
-    private static final Integer[] ARR_INDEXES = { 5, 6, 7, 8, 1};
-    private static final Set<Integer> RANDOM_INDEXES = new HashSet<>(Arrays.asList(ARR_INDEXES));
     private static final ReflectUi UI = new ReflectUi();
     // This attribute should be set as static to avoid confusion if a new object is created.
     // It means exit from self reflection back to main interface
