+++ src/main/java/wellnus/reflection/feature/ReflectionQuestion.java
@@ -1,4 +1,4 @@
-package wellnus.reflection;
+package wellnus.reflection.feature;
 
 //@@author wenxin-c
 /**
