+++ src/test/java/wellnus/reflection/FavoriteCommandTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 
 import wellnus.command.CommandParser;
 import wellnus.exception.BadCommandException;
+import wellnus.reflection.command.LikeCommand;
+import wellnus.reflection.feature.QuestionList;
 
 class FavoriteCommandTest {
     private static final String LIKE_COMMAND = "like 1";
