+++ src/test/java/wellnus/reflection/LikeCommandTest.java
@@ -12,6 +12,8 @@ import org.junit.jupiter.api.Test;
 
 import wellnus.command.CommandParser;
 import wellnus.exception.BadCommandException;
+import wellnus.reflection.command.LikeCommand;
+import wellnus.reflection.feature.QuestionList;
 
 // @@author wenxin-c
 class LikeCommandTest {
