+++ src/test/java/wellnus/reflection/QuestionListTest.java
@@ -4,6 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.ArrayList;
 
 import org.junit.jupiter.api.Test;
+import wellnus.reflection.feature.QuestionList;
+import wellnus.reflection.feature.ReflectionQuestion;
 
 // @@author wenxin-c
 class QuestionListTest {
