+++ src/test/java/wellnus/reflection/ReflectionManagerTest.java
@@ -8,6 +8,7 @@ import java.util.HashMap;
 import org.junit.jupiter.api.Test;
 
 import wellnus.exception.BadCommandException;
+import wellnus.reflection.feature.ReflectionManager;
 
 class ReflectionManagerTest {
     private static final String EMPTY_STRING = "";
