+++ src/test/java/wellnus/reflection/ReflectionQuestionTest.java
@@ -7,6 +7,7 @@ import java.io.PrintStream;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import wellnus.reflection.feature.ReflectionQuestion;
 
 class ReflectionQuestionTest {
     private static final String QUESTION = "How's today?";
